fix: second-pass review fixes for the ADAPT sample floor
- floor compares as numeric (floor(...)::int raised on 1e30)
- migration 1806 treats a JSON-null key as absent and holds off
  trigger_mark_stale_on_config_update (372 results were marked stale in a dry run)
- golden test clears the env knob at module level (beforeAll ran after collection)
- env test stubs dotenv so re-imports cannot re-read apps/worker/.env
- stale test comment about a floor of 0

+++ apps/worker/src/pipelines/helpers/adapt/results/sql-fragments.ts
@@ -49,7 +49,8 @@ export class AdaptSQLFragments {
    * The config value is untrusted (`config_data` is stored from the API body under a bare
    * `@IsObject()`): a non-number is ignored rather than aborting the whole batch with a
    * cast error, a fraction is floored, and `GREATEST(1, …)` keeps a zero or negative
-   * override from switching the floor off. Both `count` columns are nullable in the
+   * override from switching the floor off. The comparison stays numeric: an `::int` on
+   * the configured value raised `integer out of range` on 1e30. Both `count` columns are nullable in the
    * entities, and a NULL here must read as "no samples", not as a NULL `control_exists`
    * that would skip the `incomparable` branch and fall through to `no difference`.
    */
@@ -57,7 +58,7 @@ export class AdaptSQLFragments {
     const configured = "wcc.compare_config->'thresholds'->'minSampleCount'";
     const floor =
       `GREATEST(1, COALESCE(CASE WHEN jsonb_typeof(${configured}) = 'number' ` +
-      `THEN floor((${configured})::text::numeric)::int END, ${adaptMinSampleCount()}))`;
+      `THEN floor((${configured})::text::numeric) END, ${adaptMinSampleCount()}))`;
     return `wcc.control_row_exists
                 AND COALESCE(wcc.test_n, 0) >= ${floor}
                 AND COALESCE(wcc.control_n, 0) >= ${floor} as control_exists`;
